move case search tests to case search app
