+++ corehq/apps/case_search/tests/test_case_search_es_queries.py
@@ -1,0 +1,252 @@
+from collections import OrderedDict
+
+from django.test import TestCase
+
+from corehq.apps.case_search.const import CASE_SEARCH_MAX_RESULTS
+from corehq.apps.case_search.models import CaseSearchConfig, IgnorePatterns
+from corehq.apps.case_search.tests.utils import get_case_search_query
+from corehq.apps.es.tests.utils import ElasticTestMixin, es_test
+
+DOMAIN = 'mighty-search'
+
+
+@es_test
+class CaseSearchTests(ElasticTestMixin, TestCase):
+    def setUp(self):
+        super(CaseSearchTests, self).setUp()
+        self.config, created = CaseSearchConfig.objects.get_or_create(pk=DOMAIN, enabled=True)
+
+    def test_add_blacklisted_ids(self):
+        criteria = {
+            "commcare_blacklisted_owner_ids": "id1 id2 id3,id4"
+        }
+        expected = {
+            "query": {
+                "bool": {
+                    "filter": [
+                        {'terms': {'domain.exact': [DOMAIN]}},
+                        {"terms": {"type.exact": ["case_type"]}},
+                        {"term": {"closed": False}},
+                        {
+                            "bool": {
+                                "must_not": {
+                                    "terms": {
+                                        "owner_id": [
+                                            "id1",
+                                            "id2",
+                                            "id3,id4"
+                                        ]
+                                    }
+                                }
+                            }
+                        },
+                        {"match_all": {}}
+                    ],
+                    "must": {
+                        "match_all": {}
+                    }
+                }
+            },
+            "sort": [
+                "_score",
+                "_doc"
+            ],
+            "size": CASE_SEARCH_MAX_RESULTS
+        }
+
+        self.checkQuery(
+            get_case_search_query(DOMAIN, ['case_type'], criteria),
+            expected
+        )
+
+    def test_add_ignore_pattern_queries(self):
+        rc = IgnorePatterns(
+            domain=DOMAIN,
+            case_type='case_type',
+            case_property='name',
+            regex=' word',
+        )                       # remove ' word' from the name case property
+        rc.save()
+        self.config.ignore_patterns.add(rc)
+        rc = IgnorePatterns(
+            domain=DOMAIN,
+            case_type='case_type',
+            case_property='name',
+            regex=' gone',
+        )                       # remove ' gone' from the name case property
+        rc.save()
+        self.config.ignore_patterns.add(rc)
+        rc = IgnorePatterns(
+            domain=DOMAIN,
+            case_type='case_type',
+            case_property='special_id',
+            regex='-',
+        )                       # remove '-' from the special id case property
+        rc.save()
+        self.config.ignore_patterns.add(rc)
+        self.config.save()
+        rc = IgnorePatterns(
+            domain=DOMAIN,
+            case_type='case_type',
+            case_property='phone_number',
+            regex='+',
+        )                       # remove '+' from the phone_number case property
+        rc.save()
+        self.config.ignore_patterns.add(rc)
+        self.config.save()
+
+        criteria = OrderedDict([
+            ('phone_number', '+91999'),
+            ('special_id', 'abc-123-546'),
+            ('name', "this word should be gone"),
+            ('other_name', "this word should not be gone"),
+        ])
+
+        expected = {
+            "query": {
+                "bool": {
+                    "filter": [
+                        {'terms': {'domain.exact': [DOMAIN]}},
+                        {"terms": {"type.exact": ["case_type"]}},
+                        {"term": {"closed": False}},
+                        {"match_all": {}}
+                    ],
+                    "must": {
+                        "bool": {
+                            "must": [
+                                {
+                                    "nested": {
+                                        "path": "case_properties",
+                                        "query": {
+                                            "bool": {
+                                                "filter": [
+                                                    {
+                                                        "bool": {
+                                                            "filter": [
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.key.exact": "phone_number"
+                                                                    }
+                                                                },
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.value.exact": "91999"
+                                                                    }
+                                                                }
+                                                            ]
+                                                        }
+                                                    }
+                                                ],
+                                                "must": {
+                                                    "match_all": {}
+                                                }
+                                            }
+                                        }
+                                    }
+                                },
+                                {
+                                    "nested": {
+                                        "path": "case_properties",
+                                        "query": {
+                                            "bool": {
+                                                "filter": [
+                                                    {
+                                                        "bool": {
+                                                            "filter": [
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.key.exact": "special_id"
+                                                                    }
+                                                                },
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.value.exact": "abc123546"
+                                                                    }
+                                                                }
+                                                            ]
+                                                        }
+                                                    }
+                                                ],
+                                                "must": {
+                                                    "match_all": {}
+                                                }
+                                            }
+                                        }
+                                    }
+                                },
+                                {
+                                    "nested": {
+                                        "path": "case_properties",
+                                        "query": {
+                                            "bool": {
+                                                "filter": [
+                                                    {
+                                                        "bool": {
+                                                            "filter": [
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.key.exact": "name"
+                                                                    }
+                                                                },
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.value.exact": "this should be"
+                                                                    }
+                                                                }
+                                                            ]
+                                                        }
+                                                    }
+                                                ],
+                                                "must": {
+                                                    "match_all": {}
+                                                }
+                                            }
+                                        }
+                                    }
+                                },
+                                {
+                                    "nested": {
+                                        "path": "case_properties",
+                                        "query": {
+                                            "bool": {
+                                                "filter": [
+                                                    {
+                                                        "bool": {
+                                                            "filter": [
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.key.exact": "other_name"
+                                                                    }
+                                                                },
+                                                                {
+                                                                    "term": {
+                                                                        "case_properties.value.exact": "this word should not be gone"
+                                                                    }
+                                                                }
+                                                            ]
+                                                        }
+                                                    }
+                                                ],
+                                                "must": {
+                                                    "match_all": {}
+                                                }
+                                            }
+                                        }
+                                    }
+                                }
+                            ]
+                        }
+                    }
+                }
+            },
+            "sort": [
+                "_score",
+                "_doc"
+            ],
+            "size": CASE_SEARCH_MAX_RESULTS
+        }
+        self.checkQuery(
+            get_case_search_query(DOMAIN, ['case_type'], criteria),
+            expected,
+            validate_query=False
+        )
